+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [economy] feature: improvements to this module
+  - feature: new `%fuji:balance <currency-id>` placeholder, to display the `balance` of a specified `currency` for the player.
   - feature: improve the `error feedback` when failed to get specified `economy account`. 
   - feature: introduce the `--confirm` optional argument for `/economy clear` command.